+++ frontend/components/NavDrawerCategoriesTree/components/CategoriesItem/index.jsx
@@ -1,12 +1,9 @@
-import React, {
-  useState, useCallback, useMemo, useEffect,
-} from 'react';
+import React, { useState, useCallback, useMemo, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { css } from 'glamor';
 import classNames from 'classnames';
-import { ChevronIcon } from '@shopgate/engage/components';
 import { themeConfig } from '@shopgate/pwa-common/helpers/config';
-import { HtmlSanitizer } from '@shopgate/engage/components';
+import { HtmlSanitizer, ChevronIcon } from '@shopgate/engage/components';
 import CategoriesItemChildren from '../CategoriesItemChildren';
 import Item from '../Item';
 import connect from './connector';
@@ -38,7 +35,7 @@ const styles = {
     transform: 'rotateZ(-270deg)',
   }).toString(),
   open: css({
-    //  borderRight: '2px solid var(--color-primary)',
+    borderRight: '2px solid var(--color-primary)',
   }).toString(),
   transitionBlock: css({
     transition: `max-height ${animationDuration}ms cubic-bezier(0, 1, 0, 1)`,
@@ -49,35 +46,23 @@ const styles = {
     transition: `max-height ${animationDuration * 2}ms ease-in-out !important`,
     maxHeight: '1000vh !important',
   }).toString(),
+  drawer: css({
+    '&:not(:empty)': {
+      // Padding is only applied when HTMLSanitizer renders content
+      width: '100%',
+      padding: '0px 10px 25px 10px',
+    },
+    ' img': {
+      marginTop: '10px',
+    },
+    ' p': {
+      margin: '5px 0px 5px 0px',
+    },
+    ' a': {
+      textDecoration: 'underline',
+    },
+  }).toString(),
 };
-
-
-// Styling for Content
-
-css.global('.category_drawer__content',
-  { 
-    width: '100%',
-    padding: '0px 10px 25px 10px',
-  }
-);
-
-css.global('.category_drawer__content img',
-  { 
-    marginTop: '10px',
-  }
-);
-
-css.global('.category_drawer__content p',
-  { 
-    margin: '5px 0px 5px 0px',
-  }
-);
-
-css.global('.category_drawer__content a',
-  { 
-    textDecoration: 'underline',
-  }
-);
 
 /**
  * The CategoriesItem component
@@ -92,7 +77,6 @@ const CategoriesItem = ({
   content,
 }) => {
   const [isOpen, setIsOpen] = useState(false);
-  const [isPageSwitcherInitalState, setisPageSwitcherInitalState] = useState(true);
 
   const {
     maxCategoryNesting,
@@ -110,6 +94,12 @@ const CategoriesItem = ({
     setIsOpen(isPartOfCategoryPath);
   }, [isPartOfCategoryPath]);
 
+  useEffect(() => {
+    if (pageSwitcher === true) {
+      setIsOpen(true);
+    }
+  }, [pageSwitcher]);
+
   const hasSubcategories = useMemo(() => category && category.childrenCount !== 0, [category]);
   const maxNestingReached = useMemo(() => level + 1 === maxCategoryNesting, [
     level,
@@ -118,12 +108,6 @@ const CategoriesItem = ({
 
   const handleClick = useCallback(() => {
     setIsOpen(!isOpen);
-    // ###BOA### 
-    // Logic for page switcher (if used with @shopgate-project/page-switcher)
-    if (level === 0) {
-      setisPageSwitcherInitalState(false);
-    }
-    // // ###EOA### 
   }, [isOpen]);
 
   const handleOpenCategory = useCallback(() => {
@@ -131,12 +115,6 @@ const CategoriesItem = ({
   }, [categoryId, openCategory]);
 
   const buttonRight = useMemo(() => {
-    // ###BOA### 
-    // Logic for page switcher (if used with @shopgate-project/page-switcher)
-    if (pageSwitcher && isPageSwitcherInitalState && level === 0) {
-      setIsOpen(true);
-    }
-    // ###EOA###
     if (!maxNestingReached && hasSubcategories) {
       return (
         <button type="button" onClick={handleClick} className={styles.chevronButton}>
@@ -171,7 +149,7 @@ const CategoriesItem = ({
       { !maxNestingReached && hasSubcategories && subcategories && (
         <div className={classes}>
           <CategoriesItemChildren subcategories={subcategories} level={level + 1} />
-          <HtmlSanitizer className="category_drawer__content">
+          <HtmlSanitizer className={styles.drawer}>
             {content}
           </HtmlSanitizer>
         </div>
@@ -183,8 +161,9 @@ const CategoriesItem = ({
 CategoriesItem.propTypes = {
   category: PropTypes.shape(),
   categoryId: PropTypes.string,
-  content: PropTypes.oneOfType([PropTypes.array, PropTypes.object]),
+  content: PropTypes.arrayOf(PropTypes.string),
   level: PropTypes.number,
+  pageSwitcher: PropTypes.bool,
   subcategories: PropTypes.arrayOf(PropTypes.shape()),
 };
 
@@ -194,6 +173,7 @@ CategoriesItem.defaultProps = {
   categoryId: null,
   content: null,
   level: 0,
+  pageSwitcher: false,
 };
 
 export default connect(CategoriesItem);